script/requirements.txt
google-generativeai
python-dotenv
article-parser
requests